move SSE code into stream.ts
