+++ src/index.ts
@@ -16,6 +16,7 @@ import { Event } from './types'
 import { defaultOptions, defer, logError, MIN_EVENTS_SYNC_INTERVAL } from './utils'
 import { loadFromCache, removeCachedEvaluation, saveToCache, updateCachedEvaluation } from './cache'
 import { addMiddlewareToFetch } from './request'
+import { streamer } from "./stream";
 
 const SDK_VERSION = '1.15.0'
 const SDK_INFO = `Javascript ${SDK_VERSION} Client`
@@ -49,7 +50,6 @@ const convertValue = (evaluation: Evaluation) => {
 }
 
 const initialize = (apiKey: string, target: Target, options?: Options): Result => {
-  const SSE_TIMEOUT_MS = 30000;
 
   let closed = false
   let environment: string
@@ -420,11 +420,6 @@ const initialize = (apiKey: string, target: Target, options?: Options): Result =
   }
 
   const startStream = () => {
-    // TODO: Implement polling when stream is disabled
-    if (!configurations.streamEnabled) {
-      logDebug('Stream is disabled by configuration. Note: Polling is not yet supported')
-      return
-    }
 
     const handleFlagEvent = (event: StreamEvent): void => {
       switch (event.event) {
@@ -487,105 +482,16 @@ const initialize = (apiKey: string, target: Target, options?: Options): Result =
       }
     }
 
-    const processData = (data: any): void => {
-      data.toString().split(/\r?\n/).forEach(processLine);
-    }
-
-    const processLine = (line: string): void => {
-      if (line.startsWith('data:')) {
-        const event: StreamEvent = JSON.parse(line.substring(5))
-        logDebug('Received event from stream: ', event)
-
-        if (event.domain === 'flag') {
-          handleFlagEvent(event)
-        } else if (event.domain === 'target-segment') {
-          handleSegmentEvent(event)
-        }
-      }
-    }
-
-    const getRandom = (min, max) => {
-      return Math.round(Math.random() * (max - min) + min)
-    }
-
-    const onConnected = () => {
-      logDebug('Stream connected')
-      eventBus.emit(Event.CONNECTED)
-    };
-
-    const onDisconnect = () => {
-      clearInterval(readTimeoutCheckerId)
-      const reconnectDelayMs = getRandom(1000, 10000)
-      logDebug('Stream disconnected, will reconnect in ' + reconnectDelayMs + 'ms')
-      eventBus.emit(Event.DISCONNECTED)
-      setTimeout(startStream, reconnectDelayMs)
-    };
-
-    const onFailed = (msg: string) => {
-      if (!!msg) {
-        logError('Stream has issue', msg)
-      }
-      eventBus.emit(Event.ERROR_STREAM, msg)
-      eventBus.emit(Event.ERROR, msg)
-      onDisconnect()
-    }
-
     const url = `${configurations.baseUrl}/stream?cluster=${clusterIdentifier}`
-
-    const sseHeaders = {
-      'Cache-Control': 'no-cache',
-      'Accept': 'text/event-stream',
-      'API-Key': apiKey,
-      ...standardHeaders
-    }
-
-    logDebug('SSE HTTP start request', url);
-
-    const xhr = new XMLHttpRequest();
-    xhr.open("GET", url)
-    for (const [header, value] of Object.entries(sseHeaders)) {
-      xhr.setRequestHeader(header, value)
-    }
-    xhr.timeout = 24 * 60 * 60 * 1000 // Force SSE to reconnect after 24hrs
-    xhr.onerror = () => {
-      onFailed('XMLHttpRequest error on SSE stream');
-    }
-    xhr.onabort = () => {
-      logDebug('SSE aborted');
-      onFailed(null);
-    }
-    xhr.ontimeout = () => {
-      onFailed('SSE timeout');
-    }
-    xhr.onload = () => {
-      if (xhr.status >= 400 && xhr.status <= 599) {
-        onFailed(`HTTP code ${xhr.status}`);
-        return;
+    streamer(eventBus, configurations, url, apiKey, standardHeaders, (event) => {
+      if (event.domain === 'flag') {
+        handleFlagEvent(event)
+      } else if (event.domain === 'target-segment') {
+        handleSegmentEvent(event)
       }
-      onConnected();
-    }
-
-    let offset = 0;
-    let lastActivity = Date.now()
-
-    xhr.onprogress = (event) => {
-      lastActivity = Date.now()
-      const data = xhr.responseText.slice(offset);
-      offset += data.length;
-      logDebug("SSE GOT: " + data)
-      processData(data);
-    }
-
-    const readTimeoutCheckerId = setInterval(() => {
-      // this task will kill and restart the SSE connection if no data or heartbeat has arrived in a while
-      if (lastActivity < Date.now() - SSE_TIMEOUT_MS) {
-        logDebug("SSE read timeout")
-        xhr.abort()
-      }
-    }, SSE_TIMEOUT_MS)
-
-    xhr.send()
+    })
   }
+
 
   const on: EventOnBinding = (event, callback) =>
     eventBus.on((event as unknown) as EventType, (callback as unknown) as WildcardHandler)
